+++ src/main/java/capstone/project/credit_manager/domain/accounts/Student.java
@@ -1,8 +1,8 @@
 package capstone.project.credit_manager.domain.accounts;
 
-import capstone.project.credit_manager.common.ErrorResponse;
 import capstone.project.credit_manager.domain.Department;
 import capstone.project.credit_manager.domain.StudentStatus;
+import capstone.project.credit_manager.web.exception.ErrorResponse;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -37,7 +37,7 @@ public class Student extends Account {
         super(accountId, password, major, role);
 
         if (super.getRole() != role) {
-            throw new FailedCreateAccount(new ErrorResponse("NOT_MATCH_ROLE", "학생은 관리자 계정으로 가입할 수 없습니다."));
+            throw new FailedCreateAccountException(new ErrorResponse("NOT_MATCH_ROLE", "학생은 관리자 계정으로 가입할 수 없습니다."));
         }
         // TODO: 2020/06/11 위의 필드 생성자, add함수 추가
     }
