+++ src/main/java/capstone/project/credit_manager/service/exception/FailedAuthenticationException.java
@@ -1,4 +1,4 @@
-package capstone.project.credit_manager.service;
+package capstone.project.credit_manager.service.exception;
 
 public class FailedAuthenticationException extends RuntimeException {
     public FailedAuthenticationException() {
